fixed lint issues
